Adding inline images to body of blog post
Default behavior is that inline images do not appear in the body of the blog post. This code change allows for images to appear. 

Please note, there is line of CSS that needs to be commented out regarding the behavior of headlines in relation to images. I'll be addressing that in a separate pull request.